Ai_project/requirements.txt
Django==4.2.14
djangorestframework
pymongo>=4.0 
celery[redis] 
tree-sitter>=0.20.0,<0.21.0 
tree-sitter-languages>=1.6.0,<1.7.0 
tree-sitter-java 
djangorestframework-simplejwt
django-cors-headers==4.9.0
drf-yasg
django-celery-results
django-celery-beat
django-environ
setuptools
wheel
pydantic==1.10.13
gunicorn
whitenoise
requests
python-dotenv
weasyprint>=60.0
cffi>=1.15.0 
cairocffi>=1.4.0 
xhtml2pdf>=0.2.5
reportlab>=4.0.0
Pillow>=9.0.0 
html5lib>=1.1
